Add a generic "quiz this field based on some other field" functionality
Within the memory quiz subcommands, there is a common pattern of "for this type of thing X with a field value of Y, what is this other field value?" For instance, if a country struct has a name, a capital, a size rank, and a currency, we might want to quiz the name from the size or vice versa or the capital from the name or vice versa and so on.

Right now, all the memory quizzes have to implement all those queries independently. One method for name from capital, one method for capital from name, and so on. This leads to a lot of boilerplate but also means those need to actually be coded.

This new functionality lets a struct be annotated such that reflection code can generate all the possible queries.

Example
type country struct {
  sizeRank int `crossquery:"all"`
  name string `crossquery:"all"`
  capital string `crossquery:"all"`
  currency string `crossquery:"guess"`
}

This means that the code could generate quizzes such as name from size rank, size rank from name, capital from name, and so on. Flagging a field as "guess" means that it will only show up as something you need to guess. It doesn't make sense to ask which countries uses "franc" as a currency because there are many possible answers. Flagging it as "guess" means you'll only ever be asked to name the currency for a given country, size, capital, etc. and not the reverse. (There is an equivalent "given" for completeness sake)