Skip empty fields when collecting cross-query fields
This allows handling the case where most entities have some value for
the field, but some don't. For instance, Pennsylvania doesn't have  a
state bird, which we currently handle poorly by setting to None.

Also, this allows fields "in development" (set for some entities in a
list but not all) to be cross-queried right away, even if the full list
of entities doesn't yet have a value for the given field.