Add an awk file that can analyze spelling_bee.go for "triplets"
Triplets are my attempt to more actively be able to finish Spelling Bee
each day. Right now, the strategy is somewhat passive. If I see
one member of a word family, I can access all the others and type them
in. But looking at what trigrams are in each family lets me more
actively hunt down the families that might be there. For instance, AIR
as a triplet in the grid _could_ mean that FAIR, etc. are also there. It
could also mean that TIARA, etc. are there, and so on